Dev (#10)
* fixing assembly version

* Only add default region when no region is set otherwise leave the region that was set.

* adding import options for translated xml using nswdb xml format.

* version bump

+++ titledbConverter/Services/NswReleaseService.cs
@@ -19,15 +19,18 @@ public class NswReleaseService(SqliteDbContext dbContext) : INswReleaseService
     /// </summary>
     /// <param name="xmlFilePath">Path to the XML file containing release information</param>
     /// <returns>Number of records successfully imported</returns>
-    public async Task<int> ImportReleasesFromXmlAsync(string xmlFilePath)
+    public async Task<int> ImportReleasesFromXmlAsync(string xmlFilePath, bool overwrite = false)
     {
         try
         {
             var doc = XDocument.Load(xmlFilePath);
             var releases = ParseReleases(doc);
-            
-            await dbContext.Database.ExecuteSqlAsync($"DELETE FROM NswReleaseTitles");
-            await dbContext.Database.ExecuteSqlAsync($"DELETE FROM sqlite_sequence WHERE name = 'NswReleaseTitles'");
+            if (overwrite)
+            {
+                await dbContext.Database.ExecuteSqlAsync($"DELETE FROM NswReleaseTitles");
+                await dbContext.Database.ExecuteSqlAsync($"DELETE FROM sqlite_sequence WHERE name = 'NswReleaseTitles'");
+            }
+
             await dbContext.BulkInsertAsync(releases);
 
             return releases.Count;
@@ -37,6 +40,25 @@ public class NswReleaseService(SqliteDbContext dbContext) : INswReleaseService
             Console.WriteLine($"Error importing releases: {ex.Message}");
             throw;
         }
+    }
+
+    public async Task<int> ImportReleasesFromDirectoryAsync(string directoryPath)
+    {
+        var xmlFiles = Directory.GetFiles(directoryPath, "*.xml");
+        var xmlFileRecordCount = 0;
+        foreach (var filePath in xmlFiles)
+        {
+            try
+            {
+                xmlFileRecordCount += await ImportReleasesFromXmlAsync(filePath);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error importing releases from {filePath}: {ex.Message}");
+            }
+        }
+
+        return xmlFileRecordCount;
     }
 
     /// <summary>
@@ -81,22 +103,9 @@ public class NswReleaseService(SqliteDbContext dbContext) : INswReleaseService
                 titleId);
             var (cleanTitle, revision) = TitleParser.ExtractTitleAndRevision(
                 titleName);
-            var elementId = 0;
-            if (!string.IsNullOrEmpty(GetElementValue(releaseElement, "id")))
-            {
-                elementId = int.Parse(GetElementValue(releaseElement, "id"));
-            }
-            
-            if (!string.IsNullOrEmpty(GetElementValue(releaseElement, "version")))
-            {
-                var xmlVersion = GetElementValue(releaseElement, "version");
-                var numericVersion = Regex.Replace(xmlVersion, @"[^\d]", "");
-                version = uint.Parse(numericVersion);
-            }
-
             var release = new NswReleaseTitle
             {
-                Id = elementId,
+                //Id = int.Parse(GetElementValue(releaseElement, "id")),
                 ApplicationId = applicationId,
                 TitleName = cleanTitle,
                 Revision = revision,
